Struct r1_4_documentation::Person

pub struct Person {
    pub(crate) name: String,
}

A human being is represented here

Fields
name: String

A person must have a name, no matter how much Juliet may hate it

Implementations

impl Person

pub fn new(name: &str) -> Person

Creates a person with the given name.

Examples

//You can have rust code between fences inside the comments
//If you pass --test to `rustdoc`, it will even test if for you!
use r1_4_documentation::Person;
let person = Person::new("name");

pub fn hello(&self)

Gives a friendly hello!

Says “Hello, name” to the Person it is called on